Use uint64 for VU IDs

+++ lib/executor/ramping_arrival_rate_test.go
@@ -184,7 +184,7 @@ func TestRampingArrivalRateRunUnplannedVUs(t *testing.T) {
 		time.Sleep(time.Millisecond * 200)
 		cur = atomic.LoadInt64(&count)
 		require.NotEqual(t, cur, int64(2))
-		return runner.NewVU(int64(es.GetUniqueVUIdentifier()), engineOut)
+		return runner.NewVU(es.GetUniqueVUIdentifier(), engineOut)
 	})
 	err = executor.Run(ctx, engineOut)
 	assert.NoError(t, err)
@@ -234,7 +234,7 @@ func TestRampingArrivalRateRunCorrectRateWithSlowRate(t *testing.T) {
 		cur = atomic.LoadInt64(&count)
 		require.NotEqual(t, cur, int64(1))
 
-		return runner.NewVU(int64(es.GetUniqueVUIdentifier()), engineOut)
+		return runner.NewVU(es.GetUniqueVUIdentifier(), engineOut)
 	})
 	err = executor.Run(ctx, engineOut)
 	assert.NoError(t, err)
